Clarify in man page that pipewire-pulse runs its own pipewire process
Prior to this commit, the man page for pipewire-pulse did not
make it clear that although pipewire-pulse is "integrated"
in a general sense with the pipewire system, it runs its own
pipewire process, that is, in some sense, independent. For
example, restarting the pipewire service does not restart
the pipewire-pulse service.

The aim of this commit is to improve people's understanding of the
relation between pipewire-pulse and pipewire in the most obvious
place where they would expect to find it.