pipewire: log: remove `_pw_log_topic_new()`
This function was introduced in 52bd80aaa45c8b ("log: add topic loggers and a default topic")
but that commit made no use of it, and no subsequent commits ever
touched the function in any way.

No code in the repository, on debian codesearch, or on github uses it.
So remove it.