acp: Add idxset comparison and reverse search operations
Add contains() function that checks if an item is in an idxset. Add
isdisjoint(), issubset(), issuperset() and equals() functions that
element-wise compare two idxsets.

Add complementary functions to the existing idxset search(), first(),
next() functions that work in the reverse direction: reverse_search(),
last() and previous().

Add steal_last() to remove and return the last item in the idxset.

[Alper: original patches for PulseAudio, split from other commit]
Co-developed-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>
Link: https://gitlab.freedesktop.org/pulseaudio/pulseaudio/-/commit/fb63e589310fab20e60c46bb40c7b7acab5eeac9
Link: https://gitlab.freedesktop.org/pulseaudio/pulseaudio/-/commit/ec668ac44bc6e666123f22f2696745dcdce98fed
Link: https://gitlab.freedesktop.org/pulseaudio/pulseaudio/-/commit/97d9c28579c7c7400969fd93f911e7745fb483ef