jack: improve running check
We already checked if a node is_jack when it appeared in the registry
with the client.api property, so don't check again with ALWAYS_PROCESS
when the node properties changed.

When we get a node info about our own state, we can use the active state
of the node to decide if we are running or not.

See #3794